+++ languages/pt_BR.ts
@@ -195,6 +195,14 @@
         <translation type="unfinished"></translation>
     </message>
     <message>
+        <source>Dispositivo de fones (cue)</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>Usar saída principal</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
         <location filename="../build/Desktop_Qt_6_9_1-Debug/LaraRadio/configdialog.ui" line="209"/>
         <location filename="../build/Desktop_Qt_6_9_1-Release/LaraRadio/configdialog.ui" line="209"/>
         <location filename="../configdialog.ui" line="209"/>